rockets

+++ Projectiles/Rockets/RocketIce.cs
@@ -1,16 +1,15 @@
 using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using System.Collections.Generic;
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
+using System;
 using AmmoboxPlus.NPCs;
 
 namespace AmmoboxPlus.Projectiles {
-    public class BulletIce : ModProjectile {
+    public class RocketIce : ModProjectile {
 
         public override void SetStaticDefaults() {
-            DisplayName.SetDefault("Ice Bullet");
+            DisplayName.SetDefault("Ice Rocket");
             ProjectileID.Sets.TrailCacheLength[projectile.type] = 5;
             ProjectileID.Sets.TrailingMode[projectile.type] = 0;
         }
@@ -18,20 +17,19 @@ namespace AmmoboxPlus.Projectiles {
         public override void SetDefaults() {
             projectile.width = 8;
             projectile.height = 8;
-            projectile.aiStyle = 1;
+            projectile.aiStyle = 16;
             projectile.ranged = true;
             projectile.friendly = true;
             projectile.hostile = false;
             projectile.alpha = 1;
+            projectile.timeLeft = 600;
+
             projectile.ignoreWater = true;
             projectile.tileCollide = true;
-            projectile.extraUpdates = 1;
-            projectile.scale = 2f;
-            projectile.spriteDirection = 1;
-            aiType = ProjectileID.Bullet;
         }
 
         public override void OnHitNPC(NPC target, int damage, float knockback, bool crit) {
+            //  TODO: I have to rewrite this someday
 
             //  Check if enemy is in water
             //  Horrible way to actually do it, might not get all edge cases
@@ -54,18 +52,20 @@ namespace AmmoboxPlus.Projectiles {
                         }
                         ++index;
                     }
-                } else {
+                }
+                else {
                     target.AddBuff(mod.BuffType<Buffs.Cold>(), 500);
                 }
             } else { //  No stuck limit
 
                 //  If enemy is in water, apply higher chance
                 if (Main.tile[posXlow, posYlow].liquid > 0 || Main.tile[posXhi, posYhi].liquid > 0) {
-                    if (WorldGen.genRand.Next(50) == 0) {
+                    if (WorldGen.genRand.Next(3) == 0) {
                         processAddBuffIce(ref target, mod.BuffType<Buffs.Stuck>(), 300);
                     }
                 } else {
-                    if (WorldGen.genRand.Next(100) == 0) {
+                    int temp = WorldGen.genRand.Next(6);
+                    if (temp == 0 || temp == 1 || temp == 2 || temp == 3) {
                         processAddBuffIce(ref target, mod.BuffType<Buffs.Stuck>(), 300);
                     }
                 }
@@ -86,17 +86,81 @@ namespace AmmoboxPlus.Projectiles {
             }
         }
 
-        public override bool OnTileCollide(Vector2 oldVelocity) {
-            Main.PlaySound(SoundID.Item10, projectile.position);
-            projectile.Kill();
-            return false;
-        }
-
         public override void AI() {
+            int shotFrom = projectile.GetGlobalProjectile<AmmoboxGlobalProjectile>(mod).apShotFromLauncherID;
+
+            //  Rocket launcher
+            if(shotFrom == ItemID.RocketLauncher) {
+                projectile.velocity = projectile.oldVelocity;
+                if (Math.Abs(projectile.velocity.X) < 15f && Math.Abs(projectile.velocity.Y) < 15f) projectile.velocity *= 1.1f;
+                projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
+            }
+
+            //  Grenade launcher
+            if (shotFrom == ItemID.GrenadeLauncher) {
+                //projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
+                //  If going to explode 
+                if (projectile.ai[1] == 200) {
+                    projectile.Kill();
+                } else {
+                    projectile.ai[1] += 1;
+                }
+            }
+
+            //  Proximity mine
+            if (shotFrom == ItemID.ProximityMineLauncher) {
+                if(projectile.ai[1] < 3) {
+                    projectile.velocity *= 0.98f;
+                }
+                if(projectile.ai[1] >= 3 && projectile.alpha < 150) {
+                    projectile.alpha += 1;
+                }
+
+            }
+
+            //  Snowman
+            if (shotFrom == ItemID.SnowmanCannon) {
+                projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
+                AmmoboxHelpfulMethods.chaseEnemy(projectile.identity);
+            }
+
+            //  Common for all launchers
             for (int i = 0; i < 1; i++) {
                 Dust.NewDust(projectile.position, projectile.width, projectile.height, DustID.Ice, newColor: Color.WhiteSmoke);
             }
-            Lighting.AddLight(projectile.Top, Color.SkyBlue.ToVector3());
+            Lighting.AddLight(projectile.Center, Color.SkyBlue.ToVector3());
+
+        }
+
+        public override bool OnTileCollide(Vector2 oldVelocity) {
+            int shotFrom = projectile.GetGlobalProjectile<AmmoboxGlobalProjectile>(mod).apShotFromLauncherID;
+
+            //  Rocket launcher
+            if (shotFrom ==  ItemID.RocketLauncher) {
+                projectile.Kill();
+            }
+
+            //  Proximity mine
+            if (shotFrom == ItemID.ProximityMineLauncher) {
+                if(projectile.ai[1] > 3) {
+                    projectile.velocity = Vector2.Zero;
+                } else {
+                    projectile.ai[1] += 1;
+                }
+            }
+
+            //  Snowman
+            if (shotFrom == ItemID.SnowmanCannon) {
+                projectile.Kill();
+            }
+
+            return true;
+        }
+
+        public override void Kill(int timeLeft) {
+            int shotFrom = projectile.GetGlobalProjectile<AmmoboxGlobalProjectile>(mod).apShotFromLauncherID;
+    
+            AmmoboxHelpfulMethods.explodeRocket(shotFrom, projectile.identity, largeBlast: true);
         }
 
         public void processAddBuffIce(ref NPC npc, int type, int time) {
@@ -121,5 +185,7 @@ namespace AmmoboxPlus.Projectiles {
                 npc.GetGlobalNPC<AmmoboxGlobalNPC>(mod).apStuckLimit = true;
             }
         }
+
+
     }
 }
